move fixture definition
